+++ test/basepack_test_app/app/components/some_accordion.rb
@@ -1,4 +1,11 @@
 class SomeAccordion < Netzke::Basepack::Accordion
+  # This component will be rendered immediately in the first tab
+  #
+  component :panel_zero do |c|
+    c.klass = SimplePanel
+    c.title = "Panel Zero"
+  end
+
   # This component will be dynamically loaded on expanding the second accordion pane
   component :simple_panel do |c|
     c.update_text = "Update for Panel Two"
@@ -10,6 +17,7 @@ class SomeAccordion < Netzke::Basepack::Accordion
   def configure(c)
     c.title = "Some Accordion"
     c.items = [
+      :panel_zero,
       { :html => "I'm a simple Ext.Panel", :title => "Panel One" },
       :simple_panel
     ]