+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.8.1 - WIP
 * bug fix
   * TabPanel now shows the active tab's nested component upon loading
+  * Accordion now shows the active pane's nested component upon loading
 
 # 0.8.0 - 2012-12-09
 * Basepack component renaming: